app/requirements.txt
psycopg2-binary
SQLAlchemy
InquirerPy
pandas